requirements.txt
Flask
python-dotenv
psycopg2-binary
openpyxl
pandas
